backend/requirements.txt
black==23.1.0
requests==2.26.0
Django==3.2
djoser==2.2.0
django-filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
flake8==6.0.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.10.1
Pillow==9.5.0
gunicorn==20.1.0